feat: returning results from the database

+++ theia-api/src/main/java/com/mochaeng/theia_api/query/application/web/dto/DocumentResult.java
@@ -9,7 +9,7 @@ public record DocumentResult(
     @JsonProperty("documentID") UUID documentID,
 
     String title,
-    Double similarity,
+    Float similarity,
     List<AuthorResult> authors,
 
     @JsonProperty("createdAt") LocalDateTime createdAt,
